Add Statements and Print Statement
Summary:
This is a refactor of the parser. Prior to this, parser return a
root `Expr`. Now it returns a list of `Stmt` that represent an
actual program. And then the `Intepreter` handles the list of Stmt
and executes them. In this case, the only real side effect is a
Print.

This uses `system-lambda` library to capture Stdout and assert it
is valid