Create Environment class and use in interpreter
Summary:
Lotta refactoring here. The Interpreter now is stateful with an
Environment object. ATM this represents a global environment.
Because it is stateful now, each call to `Lox::run` needs to pass
in the stateful interpreter. This allows the repl to maintain state
across invocations.

Thenwe remove the AstPrinter and the BasicParserTest because they
are no longer needed. We have ways to capture standard out in junit
and we have other ways we are testing the parser.