Add AND and OR conditional handling
Summary:
Adds a new LogicalExpr that handles the AND and OR. This expression
is shaped like BinaryExpr but behaves differently. LogicalExpr can
short circuit themselves. An Or condition can return right away
when the left side is truthy. And an AND condition can return right
away when the left side *is not* truthy.