Implement break support in a loop
Summary:
This adds a break; statement. The parser throws an error if the
break; statement appears outside of a loop by keepting a counter. A
break statement is only allowed if the counter is greater than 0.

The interpreter, when it sees a BreakStmt it modifies it's state to
indicate it should ignore the rest of the statements until the
while loop exits