Add continue statement for looping
Summary:
This adds a continue; statement to the language. The continue causes
the interpreter to ignore the rest of the loop body until the start
of the next iteration.

The parser ensures the continue statement is made inside a loop
construct and throws a parse error when it is not inside a loop