First pass at function definitions
Summary:
`LoxCallable` is the interface for things we call like functions.
`LoxFunction` is the implementation of that the interpreter can use
when given a `FunctionStmt` from the parser, it is able to construct
the `LoxFunction`, evalutate the arguments and pass them into the
`Environment`.

The parser handles this by looking for `fun` token and collects the
identifiers for the parameters and the `BlockStmt` for the function.

A `CallExpr` represents a call to a function in the `Environment`.
To start, there is one `global` native function called `clock` that
is present in this default `Environment`. And finally, when the
`Interpreter` stores a `FunctionStmt`, this can be called with a
`CallExpr`.