Variable name index implementation
As part of the challenge of this chapter there is a question about
creating a optimization for contants. Prior to this change, any
time the code referenced a variable name, it used another constant
table slot -- of which tehre are only 256. After this change, the
same variable name uses the existing slot that was already defined
by the parser.