Add test for repeated use of variable
This test shows that the work done to meet the challenge question
kindof works. The key missing point is with numbers. When the same
number is repeated, it gets a new constant in the array each time.

This seems wrong, ATM the table.h api requires ObjString* for keys,
so we would need a way to map a number to an ObjString* to fix that.