Local variables implementation
Prior to this change, blocks were not parsed and scoping wasn't
handled. In this change, the scopes are an integer that we move
up and down when encountering a { and }. The values of each
variable is located on the stack and OP_GET_LOCAL takes the index
in the stack as a parameter to fetch it to push onto the stack.

When the } block is reached, the compiler emits N OP_POP where N
is the number of local variables in the block.