First implementation of if and else
This is a basic implementation and doesn't include boolean operators
of `or` and `and` variety. `emitJump` function puts a "sentinel" jump
offset that is 0. After calculating the statement of the then block,
the jump is patched with `short` that represents the jump.

jump if false is the op code that instructs the virtual machine to
check condition and if it is false, to jump ahead of the bytes.

In the case of no else, it will jump over a POP instruction. Seems
wasteful and an opptunity for an optimization.