Use existing instructions to implement `and` and `or`
This commit uses existing jump byte code to implement `and` and
`or. And `and`, first emits a jump if false to the end of the
block, and then parses the next. If it is true, it pops the stack
to remove the value and will then check the next. If it was false,
it will jump over the rest of the block.

Or behaves differently. If the first condition is false, it jumps
*over* a single JUMP operation. If it is true, it will move on to
that jump operation. That jump will take it into the body of the
block.