Finish closure work by uplifting values
Cool trick, when the upvalue object is about to be poped off the
stack, tell the VM to move it to the heap. It does that by putting
the value into the `closed` field of the Upvalue Struct and then
updates the `location` pointer to point to that new address.
Therefore, get and set ops codes for upvalues don't have to change.