Trigger GC on allocation of new objects based on growth
In this commit, whenever a reallocate is called with a request for
more memory, compare the new memory to nextGC to see if it should be
triggered before allocating this memory

The next GC is triggerred when the heap groups by 2 times. It moves
around as the program grows and shrinks in an attempt to balance
throughput and GC latency.